Nouvelles méthodes

+++ Framework/Database/Table/CategorizedTable.php
@@ -12,35 +12,21 @@ use Framework\database\Record;
 
 class CategorizedTable
 {
-    const TABLE = 'Categorized';
-    private string $limit = '100';
-    public function __construct(protected readonly Connexion $connexion)
-    {
-
-    }
+    use BasicTable;
+    const TABLE = 'categorized';
     private function newEntity(array $data) : Categorized
     {
         $ticket = $data[0];
         $category = $data[1];
         return new Categorized($ticket, $category);
     }
-    public function execute(string $request) : Record|bool
+    public function categoriesOf(int $ticket) : array|bool
     {
-        $request .= "LIMIT $this->limit";
-        $query = $this->connexion->prepare($request);
-        if(!$query->execute())
+        if ($record = $this->select($ticket) === FALSE)
         {
-            return false;
+            return FALSE;
         }
-        if(true) {
-            $record = new Record();
-            while($datum = $query->fetch(PDO::FETCH_ASSOC))
-            {
-                $record->addDatum($this->newEntity($datum));
-            }
-            return $record;
-        }
-        return true;
+        return $record->getData();
     }
     public function select(int $ticket = null, int $category = null) : Record|null
     {
@@ -51,11 +37,11 @@ class CategorizedTable
         }
         elseif (isset($ticket))
         {
-            $request .= "WHERE ID = $ticket";
+            $request .= "WHERE ticket = $ticket";
         }
         elseif (isset($category))
         {
-            $request .= "WHERE ID = $category";
+            $request .= "WHERE category = $category";
         }
         $request .= "LIMIT $this->limit";
         return $this->execute($request);
@@ -77,8 +63,12 @@ class CategorizedTable
         }
         return TRUE;
     }
-    public function delete(int ...$IDs) : bool
+    public function delete(string $key, int ...$IDs) : bool
     {
+        if ($key !== 'ticket' && $key !== 'category')
+        {
+            return FALSE;
+        }
         foreach($IDs as $ID)
         {
             $request = 'DELETE FROM' . self::TABLE . "WHERE ID = $ID";
@@ -88,16 +78,5 @@ class CategorizedTable
             }
         }
         return TRUE;
-    }
-
-    public function getLimit(): string
-    {
-        return $this->limit;
-    }
-
-    public function setLimit(string $limit): Table
-    {
-        $this->limit = $limit;
-        return $this;
     }
 }
